backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pypdf2==3.0.1
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.6.1
anthropic==0.8.1
pandas==2.1.4
geopy==2.4.1
python-multipart==0.0.6
aiofiles==23.2.1
schedule==1.2.0
notion-client==2.2.0